requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
Jinja2==3.0.1
pymongo==3.12.0
pytango==9.3.0
pytest==4.6.9
requests==2.22.0
urllib3==1.25.8
Werkzeug==2.0.1

